// crud-angular/src/app/core/application/dtos/user.dto.ts
export interface UserDto {
  id: string;
  name: string;
  email: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateUserDto {
  name: string;
  email: string;
}

export interface UpdateUserDto {
  name?: string;
  email?: string;
  isActive?: boolean;
}